New Update v 5.3 - user UI - top chart page - top chart pane - style.css
