+++ src/style_css/style.css
@@ -224,6 +224,20 @@
     
 }
 
+.seealllabel {
+    -fx-font-size: 15px;
+    -fx-text-fill: white;
+}
+
+.seealllabel:hover {
+    -fx-underline: true;
+}
+.seealllabel:pressed {
+    -fx-underline: true;
+    -fx-text-fill: black;
+}
+
+
 .anchorpane{
     -fx-background-color : #00a8cc;
 }
@@ -271,14 +285,14 @@
 }
 
 .buttoninbrowse{
-    -fx-background-color: #111111;
-    -fx-opacity: .8;
+    -fx-background-color : #222222;
+    -fx-opacity: .5;
     -fx-background-radius: 0em;
 }
 
 .borderlist {
-    -fx-background-color: #111111;
-    -fx-opacity: .8;
+    -fx-background-color : #27496d;
+    -fx-opacity: 0.99;
 }
 
 .buttonlist{
